Add files via upload
The objective of this project is to classify whether someone has diabetes or not. So, the high-level question is “Can we be able to predict if someone suffers from diabetes before diagnosing her/him?” Dataset consists of several Medical Variables (Independent) and one Outcome Variable (Dependent). The independent variables in this dataset are: ’Pregnancies’, ’Glucose’, ’Blood Pressure’, ’Skin Thickness’, ’Insulin’, ’BMI’, ’Diabetes Pedigree Function’, ’Age’ The outcome variable value is either 1 or 0 indicating whether a person has diabetes (1) or not (0).